requirements.txt
fastapi
slowapi
uvicorn[standard]
httpx
anthropic
beautifulsoup4
python-multipart
python-dotenv
lxml
